test: isolate LoRA runtime preflight receipts
